+++ src/Configuration.hpp
@@ -20,7 +20,7 @@
 
 #pragma once
 
-static constexpr const auto k_testmap_filename = "demo-map3.tmx";
+static constexpr const auto k_testmap_filename = "demo-map2.tmx";
 static constexpr const bool
     k_region_axis_container_report_maximum_sort_and_sweep = false;
 constexpr const bool k_promised_files_take_at_least_one_frame = true;
